Switched to evaluation in floating point numbers.

+++ src/ChessEngineGreedy/engine.h
@@ -29,7 +29,7 @@ public:
   bool Start(const IMachine& position, int depth, int timeout) override;
   void Stop() override;
   void ProcessInfo(IInfoCall* cb) override;
-  int EvalPosition(const IMachine& position) const override;
+  float EvalPosition(const IMachine& position) const override;
 
 private:
   void SearchDepth(int depth) override;
@@ -37,17 +37,17 @@ private:
   void NodesPerSecond(int nps) override;
   void ReadyOk() override;
   void BestMove(const char* notation) override;
-  void BestScore(int score) override;
+  void BestScore(float score) override;
 
   void ThreadFun(boost::shared_ptr<IMachine> machine);
-  int Search(boost::shared_ptr<IMachine> machine, Set set, int depth);
+  float Search(boost::shared_ptr<IMachine> machine, Set set, int depth);
   std::vector<Move> EmunMoves() const;
   
   Set xSet(Set set) const { return set == Set::white ? Set::black : (set == Set::black ? Set::white : Set::unknown); }
-  int EvalPosition(const IMachine & position, Set set) const;
-  int EvalSide(const IMachine & position, Set set, const std::vector<Piece>& white, const std::vector<Piece>& black) const;
-  int PieceWeight(Type type) const;
-  int PositionWeight(Set set, const Piece& piece, const std::vector<Piece>& white, const std::vector<Piece>& black) const;
+  float EvalPosition(const IMachine & position, Set set) const;
+  float EvalSide(const IMachine & position, Set set, const std::vector<Piece>& white, const std::vector<Piece>& black) const;
+  float PieceWeight(Type type) const;
+  float PositionWeight(Set set, const Piece& piece, const std::vector<Piece>& white, const std::vector<Piece>& black) const;
 
   boost::asio::io_service service;
   boost::thread thread;
