+++ src/chessmachine/chessmachine.h
@@ -63,7 +63,7 @@ namespace Chai {
       // Messages sent after the search
       virtual void ReadyOk() = 0;
       virtual void BestMove(const char* notation) = 0;
-      virtual void BestScore(int score) = 0;
+      virtual void BestScore(float score) = 0; // in pawns
     };
 
     class IEngine {
@@ -81,7 +81,7 @@ namespace Chai {
       virtual bool Start(const IMachine& position, int depth, int timeout = 0) = 0;
       virtual void Stop() = 0;
       virtual void ProcessInfo(IInfoCall* cb) = 0;
-      virtual int EvalPosition(const IMachine& position) const = 0; // Evaluation of the current position. It's positive for white set of and it's negative for black.
+      virtual float EvalPosition(const IMachine& position) const = 0; // Evaluation of the current position. It's positive for white set of and it's negative for black.
 
       virtual ~IEngine() {}
     };
